CSW: Fix backtrack to CarbonSettings Fragment
The current handling results in the followup activity to ignore it for
backtracks which means when going from Location -> CarbonSettings -> Fingerprint
the move backwards from Fingerprint leads up in Location and not the expected
CarbonSettings.

The custom handling is necessary to process the result in the Google SUW,
so it has been fixed to properly work now.

Change-Id: Ifae5dc8fda07b61bd0271da1d1589059ab40cd39