CSW: Fix backtrack to CarbonSettings Fragment
The non standard handling results in the followup activity to ignore it for
backtracks which means when going from Location -> CarbonSettings -> Fingerprint
the move backwards from Fingerprint leads up in Location and not theexpected
CarbonSettings.

As the custom handling seems unecessary and the result code is not processed
we can remove it and fallback to the default BaseSetupWizardActivity handling.

Change-Id: Ifae5dc8fda07b61bd0271da1d1589059ab40cd39